+++ CHANGELOG.md
@@ -6,6 +6,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 ### Added
+ * `MysqlDatabase` `MysqlUser` Add delete policy
 ### Changed
 ### Removed
 ### Fixed